requirements.txt
python-dotenv
langchain
langchain-openai
langchain-community
pypdf
faiss-cpu